Structure Start
Begins to port star_prerouting to rust by creating the rust file, and performing the set up so that the python file can correctly call it without any compiler error.

Uses `SabreDAG` to create the dag in rust to work with, and that will be where logic for star_prerouting should mostly interact with.